refactor(test): migrate test_analyzer unique tests to test_ai_analyzer
迁移 2 个唯一测试（并发处理和顺序校验）至 test_ai_analyzer.py，删除旧文件
